i want to use axios and React Query to call from a fake API. Through this fake API, I can get prediction, get recommendation through filter.
In Map: Show only filter by average Price.

In Prediction: Longitude and Latitude auto get through full Address (if you cannot, keep longitude and latitiude)

In Introduction, have a part of feature (as dropdown menu).

In Recommendation, when View Details we move to a contact page with salesman information (email, ...).